+++ src/Meshing/MeshingGui.cpp
@@ -42,7 +42,7 @@ void MeshingGui::imgui_window(Cool::MeshExportSettings& mesh_export_settings, Da
         ImGui::Separator();
         mesh_export_settings.imgui();
 
-        Cool::ImGuiExtras::before_export_button(mesh_export_settings.path);
+        Cool::ImGuiExtras::before_export_button(mesh_export_settings.path, Cool::ExportPathChecks{});
         if (ImGui::Button(Cool::icon_fmt("Export 3D Model", ICOMOON_UPLOAD2).c_str()))
         {
             gen_and_export_mesh(_main_node_id, _meshing_settings, mesh_export_settings, data_to_pass_to_shader, data_to_generate_shader_code);