+++ src/ProjectManager/Command_PackageProjectInto.h
@@ -1,11 +1,12 @@
 #pragma once
 #include "CommandCore/CommandExecutionContext_Ref.h"
+#include "SaveAsOptions.hpp"
 
 namespace Lab {
 
 struct Command_PackageProjectInto {
     std::filesystem::path folder_path{};
-    bool                  register_project_in_the_launcher{};
+    SaveAsOptions         options{};
 
     void               execute(CommandExecutionContext_Ref const& ctx) const;
     [[nodiscard]] auto to_string() const -> std::string;