+++ src/ProjectManager/SaveAsOptions.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace Lab {
+
+struct SaveAsOptions {
+    bool register_project_in_the_launcher{true};
+    bool allow_overwrite_existing_file{false};
+};
+
+} // namespace Lab